rename: 공통 뉴스 컴포넌트 파일명 및 폴더명 수정
