+++ src/pages/HomePage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Banner from '../components/home/BannerSection';
 import RankingSection from '../components/home/Ranking/RankingSection';
 import HomeHighlight from '../components/common/Highlight/HomeHighlight';
-import HomeTeamNews from '../components/common/TeamNews/HomeTeamNews';
+import HomeNews from '../components/common/News/HomeNews';
 import MatchCardSection from '../components/home/MatchCard/MatchCardSection';
 
 function HomePage() {
@@ -12,7 +12,7 @@ function HomePage() {
       <RankingSection />
       <MatchCardSection />
       <HomeHighlight />
-      <HomeTeamNews />
+      <HomeNews />
     </main>
   );
 }
